% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{byMonth}
\alias{byMonth}
\title{Returns the missing percentage/mean/max by month}
\usage{
byMonth(df, func, vint_col)
}
\arguments{
\item{df}{Data frame with variables to check for stability}

\item{func}{Function to apply to each variable each month and compare across}

\item{vint_col}{Column that contains month}
}
\value{
Data frame with results by month and standard deviation between months
}
\description{
Returns the missing percentage/mean/max by month
}
