% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{calculateIV}
\alias{calculateIV}
\title{IV calculation}
\usage{
calculateIV(iv_data, iv_response, iv_parallel = FALSE, cores = 1)
}
\arguments{
\item{iv_data}{Dataset containing all the variables to calculate the IV for}

\item{iv_response}{Name of the response variable (it has to be binary)}

\item{iv_parallel}{Boolean, whether to use parallel processing or no. Default TRUE}

\item{cores}{Number of cores to use if the iv_parallel flag is set to TRUE}
}
\value{
Dataset with iv values for each of the attributes in the initial dataset
}
\description{
IV calculation
}
