% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{capFloor}
\alias{capFloor}
\title{Cap floor and save the cap/floor values}
\usage{
capFloor(dat, target)
}
\arguments{
\item{dat}{Data frame with variables to cap/floor}
}
\value{
List where the first element is the capped/floored dataset, and second is the dataset with details on cap and floor values
}
\description{
Cap floor and save the cap/floor values
}
