% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{getAuc}
\alias{getAuc}
\title{Helper for AUC}
\usage{
getAuc(scr, data, target)
}
\arguments{
\item{scr}{Model score}

\item{data}{Dataset containing the target variable}

\item{target}{Column name of the target variable}
}
\value{
AUC of the model score
}
\description{
Helper for AUC
}
