% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{getFScore}
\alias{getFScore}
\title{Helper for FScore}
\usage{
getFScore(scr, data, target)
}
\arguments{
\item{scr}{Model score}

\item{data}{Dataset containing the target variable}

\item{target}{Column name of the target variable}
}
\value{
FScore of the model score
}
\description{
Helper for FScore
}
