% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{getLift}
\alias{getLift}
\title{Score data - XGB models}
\usage{
getLift(model, test, target)
}
\arguments{
\item{model}{XGBoost model object}

\item{test}{Test set}

\item{target}{Column name of the target variable}
}
\value{
Lift chart for the model
}
\description{
Score data - XGB models
}
