% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{getROC}
\alias{getROC}
\title{Helper for ROC plot}
\usage{
getROC(scr, data, target)
}
\arguments{
\item{scr}{Model score}

\item{data}{Dataset containing the target variable}

\item{target}{Column name of the target variable}
}
\value{
Plot of ROC curve and the corresponding dataset
}
\description{
Helper for ROC plot
}
