% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{modelFeatures}
\alias{modelFeatures}
\title{Creates a dataset with summary of selected model features and their importance}
\usage{
modelFeatures(test, score_col = "score", features, feat_col = "feature",
  imp_col = "importance", dict, dict_feat_col = "feature",
  bureau_col = "bureau", desc_col = "description", agg = mean)
}
\arguments{
\item{test}{Scored hold-out set. Has to contain all the model variables. Currently assumed that any special values are taken care of prior to usage}

\item{score_col}{Column name of the score column}

\item{features}{Dataset containing model features ranked by importance with a corresponding importance value}

\item{feat_col}{Column name for the feature name}

\item{imp_col}{Column name for the feature importance}

\item{dict}{A dictionary}

\item{dict_feat_col}{Column name for the feature name in dictionary}

\item{bureau_col}{Column name for the bureau in the dictionary}

\item{desc_col}{Column name for the description in the dictionary}

\item{agg}{function to use as the aggregate}
}
\value{
A dataset with details about each variable, including bureau, description, and values for top and bottom 10% of the population
}
\description{
Creates a dataset with summary of selected model features and their importance
}
