% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{performQuery}
\alias{performQuery}
\title{Run a Query using RODBC}
\usage{
performQuery(query, database = "Hive")
}
\arguments{
\item{query}{String with query}

\item{database}{Database name}
}
\value{
Result of query as a data frame
}
\description{
Run a Query using RODBC
}
