% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the dataset}
\usage{
preprocess(dat, nearzero = T)
}
\arguments{
\item{dat}{Data frame to preprocess}

\item{nearzero}{Boolean, whether to remove near zero variables or not. Default TRUE}
}
\value{
Preprocessed dataset
}
\description{
Preprocess the dataset
}
