% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_helper_functions.R
\name{score}
\alias{score}
\title{Score data - caret models}
\usage{
score(model, test, target)
}
\arguments{
\item{model}{Caret model object}

\item{target}{Column name of the target variable}

\item{data}{Test set}
}
\value{
Lift chart for the model
}
\description{
Score data - caret models
}
