% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{stability_selection_lasso}
\alias{stability_selection_lasso}
\title{Runs the stability selection with lasso algorithm to select most important variables. See http://stat.ethz.ch/~nicolai/stability.pdf}
\usage{
stability_selection_lasso(data, numvars = 100, cutoff = 0.6,
  return_df = F)
}
\arguments{
\item{data}{dataset with variables and target.}

\item{numvars}{Number of variables considered in each subsample.}

\item{cutoff}{Probability cutoff, only variables with probability higher than it are selected. Can range from 0.5 to 1, recommended is from 0.6 to 0.9.}

\item{return_df}{TRUE/FALSE, whether to return just the variable list(F) or the data frame with details as well (T)}
}
\value{
Data frame consisting of selected variables and a dataframe with details if return_df == TRUE
}
\description{
Runs the stability selection with lasso algorithm to select most important variables. See http://stat.ethz.ch/~nicolai/stability.pdf
}
