% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{zeroVariance}
\alias{zeroVariance}
\title{Get names of columns with zero variance}
\usage{
zeroVariance(dat)
}
\arguments{
\item{dat}{Data frame with variables that need to be checked for zero variance}
}
\value{
Column names of variables that have zero variance
}
\description{
Get names of columns with zero variance
}
